Successfully pushed Boulder data down into the chart
